Expose template variable names in the order they're defined
This is made possible by the new ordered object decoding available
starting with Jason 1.13.

This also removes the implementation of the Collectable protocol for
Context structs (as it wasn't really that useful).

Variable order gets lost after context updates due to the current
implementation of `Enumerable`, based on Map; fixing that is left for
a future refinement (it's not too important anyway)